refactor : rename directories
