+++ apps/web/src/components/common/Avatar.tsx
